one function def per file, separate constants and types, clean up
