+++ src/types.ts
@@ -1,0 +1,7 @@
+import { RequiredOptions } from 'prettier';
+
+export interface PrettierParserOptions extends RequiredOptions {
+    importOrder: string[];
+    importOrderSeparation: boolean;
+    experimentalBabelParserPluginsList: [];
+}
