--- src/utils/constants.ts
@@ -1,6 +1,0 @@
-export const flow = 'flow';
-export const babelFlow = 'babel-flow';
-export const typescript = 'typescript';
-export const decoratorsLegacy = 'decorators-legacy';
-export const classProperties = 'classProperties';
-export const jsx = 'jsx';
