+++ src/utils/get-all-comments-from-nodes.ts
@@ -1,0 +1,12 @@
+import { Statement, CommentBlock, CommentLine } from '@babel/types';
+
+export const getAllCommentsFromNodes = (nodes: Statement[]) =>
+    nodes.reduce((acc, node) => {
+        if (
+            Array.isArray(node.leadingComments) &&
+            node.leadingComments.length > 0
+        ) {
+            acc = [...acc, ...node.leadingComments];
+        }
+        return acc;
+    }, [] as (CommentBlock | CommentLine)[]);
